finished refactoring and modularization, added necessary imports, fixed method args
